refactor: useSwap
