+++ src/utils/transactions/helpers/index.ts
@@ -1,6 +1,5 @@
 export * from './signAndSendTransaction';
 export * from './signAndSendAllTransactions';
-export * from './signAndSendTransactionsInSeries';
 export * from './signAndSendAllTransactionsInSeries';
 export * from './signAndSendTransactionsOneByOne';
 export * from './showSolanaFMLinkNotification';
